splits methods into main.rb and logic.rb
The working methods are put in the adequate file
